+++ tests/test_order_LocalQl.py
@@ -26,17 +26,43 @@ class TestLocalQl(unittest.TestCase):
 
         comp = freud.order.LocalQl(box, 1.5, 6)
 
+        with self.assertRaises(AttributeError):
+            comp.num_particles
+        with self.assertRaises(AttributeError):
+            comp.Ql
+        with self.assertRaises(AttributeError):
+            comp.ave_Ql
+        with self.assertRaises(AttributeError):
+            comp.norm_Ql
+        with self.assertRaises(AttributeError):
+            comp.ave_norm_Ql
+
         comp.compute(positions)
         npt.assert_allclose(np.average(comp.Ql), 0.57452422, rtol=1e-6)
         npt.assert_allclose(comp.Ql, comp.Ql[0], rtol=1e-6)
+
+        with self.assertRaises(AttributeError):
+            comp.ave_Ql
+        with self.assertRaises(AttributeError):
+            comp.norm_Ql
+        with self.assertRaises(AttributeError):
+            comp.ave_norm_Ql
 
         comp.computeAve(positions)
         npt.assert_allclose(np.average(comp.Ql), 0.57452422, rtol=1e-6)
         npt.assert_allclose(comp.ave_Ql, comp.ave_Ql[0], rtol=1e-6)
 
+        with self.assertRaises(AttributeError):
+            comp.norm_Ql
+        with self.assertRaises(AttributeError):
+            comp.ave_norm_Ql
+
         comp.computeNorm(positions)
         npt.assert_allclose(np.average(comp.Ql), 0.57452422, rtol=1e-6)
         npt.assert_allclose(comp.norm_Ql, comp.norm_Ql[0], rtol=1e-6)
+
+        with self.assertRaises(AttributeError):
+            comp.ave_norm_Ql
 
         comp.computeAveNorm(positions)
         npt.assert_allclose(np.average(comp.Ql), 0.57452422, rtol=1e-6)
@@ -76,17 +102,43 @@ class TestLocalQlNear(unittest.TestCase):
 
         comp = freud.order.LocalQlNear(box, 1.5, 6, 12)
 
+        with self.assertRaises(AttributeError):
+            comp.num_particles
+        with self.assertRaises(AttributeError):
+            comp.Ql
+        with self.assertRaises(AttributeError):
+            comp.ave_Ql
+        with self.assertRaises(AttributeError):
+            comp.norm_Ql
+        with self.assertRaises(AttributeError):
+            comp.ave_norm_Ql
+
         comp.compute(positions)
         npt.assert_allclose(np.average(comp.Ql), 0.57452416, rtol=1e-6)
         npt.assert_allclose(comp.Ql, comp.Ql[0], rtol=1e-6)
+
+        with self.assertRaises(AttributeError):
+            comp.ave_Ql
+        with self.assertRaises(AttributeError):
+            comp.norm_Ql
+        with self.assertRaises(AttributeError):
+            comp.ave_norm_Ql
 
         comp.computeAve(positions)
         npt.assert_allclose(np.average(comp.Ql), 0.57452416, rtol=1e-6)
         npt.assert_allclose(comp.ave_Ql, comp.ave_Ql[0], rtol=1e-6)
 
+        with self.assertRaises(AttributeError):
+            comp.norm_Ql
+        with self.assertRaises(AttributeError):
+            comp.ave_norm_Ql
+
         comp.computeNorm(positions)
         npt.assert_allclose(np.average(comp.Ql), 0.57452416, rtol=1e-6)
         npt.assert_allclose(comp.norm_Ql, comp.norm_Ql[0], rtol=1e-6)
+
+        with self.assertRaises(AttributeError):
+            comp.ave_norm_Ql
 
         comp.computeAveNorm(positions)
         npt.assert_allclose(np.average(comp.Ql), 0.57452416, rtol=1e-6)